feat(tiff): BigTIFF container write with 64-bit relocation
Widen format/tiff's copy-and-relocate serializer (relocateTIFFFromParsed,
behind tiff.Inject/tiff.InjectWithEXIF) to correctly rebuild a standalone
BigTIFF container: image-block enumeration/placeholder logic, SubIFD (0x014A)
enumeration/patching, and outer-container structural scans now dispatch on a
bigTIFF flag between classic (8-byte header/12-byte entries/4-byte fields)
and BigTIFF (16-byte header/20-byte entries/8-byte fields) widths, closing
the second half of task #264's BigTIFF write mandate.

Along the way, fixed a pre-existing bug where 0x014A SubIFD entries declared
as TIFF-Extension type 13 ("IFD") were silently corrupted because exif.Parse
follows CIPA DC-008-2023's EXIF-3.0 assignment of the SAME numeric code to
TypeUTF8 — the relocator now re-scans raw bytes for this tag instead of
trusting the exif-parsed struct.

Full support proven against 16 real BigTIFF corpus fixtures plus committed
big_cramps_{be,le}.tif; gometadata.Write's top-level BigTIFF gate (write.go)
is unchanged and remains a follow-up outside this task's declared scope.

+++ format/tiff/realfile_test.go
@@ -117,8 +117,8 @@ func TestInjectWithEXIFRealFile_PassThrough(t *testing.T) {
 
 // stripRecord records a single strip's original offset and byte count.
 type stripRecord struct {
-	srcOffset uint32
-	length    uint32
+	srcOffset uint64
+	length    uint64
 }
 
 // snapshotStrips reads the StripOffsets and StripByteCounts from e before any
@@ -230,8 +230,8 @@ func verifyImageBlocksIdentical(t *testing.T, original []byte, origStrips []stri
 			continue // zero-size strip; nothing to compare
 		}
 
-		end1 := uint64(orig.srcOffset) + uint64(orig.length)
-		end2 := uint64(resNewOff) + uint64(resSrcLen)
+		end1 := orig.srcOffset + orig.length
+		end2 := resNewOff + resSrcLen
 		if end1 > uint64(len(original)) {
 			t.Errorf("strip[%d]: original offset %d+%d out of bounds (file len %d)",
 				i, orig.srcOffset, orig.length, len(original))
